Refaktor: GraphFx-Klassen in graphing.fx verschieben
