+++ src/main/java/com/mlprograms/justmath/graphing/fx/JavaFxRuntime.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package com.mlprograms.justmath.graphfx;
+package com.mlprograms.justmath.graphing.fx;
 
 import javafx.application.Platform;
