+++ src/main/java/com/mlprograms/justmath/graphing/fx/planar/calculator/GraphFxCalculatorEngine.java
@@ -22,16 +22,16 @@
  * SOFTWARE.
  */
 
-package com.mlprograms.justmath.graphfx.planar.calculator;
+package com.mlprograms.justmath.graphing.fx.planar.calculator;
 
 import com.mlprograms.justmath.bignumber.BigNumber;
 import com.mlprograms.justmath.calculator.CalculatorEngine;
 import com.mlprograms.justmath.calculator.internal.TrigonometricMode;
-import com.mlprograms.justmath.graphfx.ReservedVariables;
-import com.mlprograms.justmath.graphfx.planar.model.PlotLine;
-import com.mlprograms.justmath.graphfx.planar.model.PlotPoint;
-import com.mlprograms.justmath.graphfx.planar.model.PlotResult;
-import com.mlprograms.justmath.graphfx.planar.view.ViewportSnapshot;
+import com.mlprograms.justmath.graphing.fx.ReservedVariables;
+import com.mlprograms.justmath.graphing.fx.planar.model.PlotLine;
+import com.mlprograms.justmath.graphing.fx.planar.model.PlotPoint;
+import com.mlprograms.justmath.graphing.fx.planar.model.PlotResult;
+import com.mlprograms.justmath.graphing.fx.planar.view.ViewportSnapshot;
 
 import java.math.MathContext;
 import java.math.RoundingMode;
