+++ src/main/java/com/mlprograms/justmath/graphing/fx/planar/view/ViewportSnapshot.java
@@ -22,17 +22,20 @@
  * SOFTWARE.
  */
 
-package com.mlprograms.justmath.graphfx.planar.model;
+package com.mlprograms.justmath.graphing.fx.planar.view;
 
 import com.mlprograms.justmath.bignumber.BigNumber;
 
 import java.util.Objects;
 
-public record PlotPoint(BigNumber x, BigNumber y) {
+public record ViewportSnapshot(BigNumber minX, BigNumber maxX, BigNumber minY, BigNumber maxY, BigNumber cellSize) {
 
-    public PlotPoint {
-        Objects.requireNonNull(x, "x must not be null");
-        Objects.requireNonNull(y, "y must not be null");
+    public ViewportSnapshot {
+        Objects.requireNonNull(minX, "minX must not be null");
+        Objects.requireNonNull(maxX, "maxX must not be null");
+        Objects.requireNonNull(minY, "minY must not be null");
+        Objects.requireNonNull(maxY, "maxY must not be null");
+        Objects.requireNonNull(cellSize, "cellSize must not be null");
     }
 
 }
