+++ src/main/java/com/mlprograms/justmath/graphing/fx/planar/view/VisibleWorldBounds.java
@@ -22,23 +22,7 @@
  * SOFTWARE.
  */
 
-package com.mlprograms.justmath.graphfx;
+package com.mlprograms.justmath.graphing.fx.planar.view;
 
-import java.util.Objects;
-
-public enum ReservedVariables {
-
-    X("x"),
-    Y("y");
-
-    private final String value;
-
-    ReservedVariables(final String value) {
-        this.value = Objects.requireNonNull(value, "value must not be null");
-    }
-
-    public String getValue() {
-        return value;
-    }
-
+record VisibleWorldBounds(double minX, double maxX, double minY, double maxY) {
 }
