+++ src/test/java/com/mlprograms/justmath/graphing/fx/JavaFxRuntimeTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2026 Max Lemberg
+ * Copyright (c) 2025-2026 Max Lemberg
  *
  * This file is part of JustMath.
  *
@@ -22,15 +22,10 @@
  * SOFTWARE.
  */
 
-package com.mlprograms.justmath.graphfx.planar.model;
+package com.mlprograms.justmath.graphing.fx;
 
-import java.util.List;
-import java.util.Objects;
+class JavaFxRuntimeTest {
 
-public record PlotLine(List<PlotPoint> plotPoints) {
-
-    public PlotLine {
-        Objects.requireNonNull(plotPoints, "plotPoints must not be null");
-    }
+    // TODO
 
 }
