+++ src/test/java/com/mlprograms/justmath/graphing/fx/planar/view/GraphFxViewerTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2026 Max Lemberg
+ * Copyright (c) 2025-2026 Max Lemberg
  *
  * This file is part of JustMath.
  *
@@ -22,20 +22,10 @@
  * SOFTWARE.
  */
 
-package com.mlprograms.justmath.graphfx.planar.view;
+package com.mlprograms.justmath.graphing.fx.planar.view;
 
-import com.mlprograms.justmath.bignumber.BigNumber;
+class GraphFxViewerTest {
 
-import java.util.Objects;
-
-public record ViewportSnapshot(BigNumber minX, BigNumber maxX, BigNumber minY, BigNumber maxY, BigNumber cellSize) {
-
-    public ViewportSnapshot {
-        Objects.requireNonNull(minX, "minX must not be null");
-        Objects.requireNonNull(maxX, "maxX must not be null");
-        Objects.requireNonNull(minY, "minY must not be null");
-        Objects.requireNonNull(maxY, "maxY must not be null");
-        Objects.requireNonNull(cellSize, "cellSize must not be null");
-    }
+    // TODO
 
 }
